refactor(main-content): dedupe, move `parseUtils`
Only a handful of the `MainContent` components are still being used, so
I moved them to where they should be — `GenericContentPage/`.
`parseUtils` were mainly being used in `TargetComponents/`, so I moved
the file there and removed `MainContent/` altogether.
