+++ src/components/TargetComponents/Evidence.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { IEvidence } from '../../models/target';
 import { OrderedList } from './Lists';
-import { parseContent } from '../MainContent/parseUtils';
+import { parseContent } from './parseUtils';
 
 export interface EvidenceProps {
   evidence: IEvidence[];
